+++ src/lib.rs
@@ -3,20 +3,16 @@
 mod char_size;
 mod draw_target;
 mod framebuf;
-mod generated;
+mod glyph_reader;
 pub mod mapping;
 mod mono_image;
 mod multi_mono_text_style;
 mod static_text;
-mod sub_image;
-
-use core::fmt;
 
 pub use char_size::CharSize;
 pub use framebuf::{BulkFlushTarget, Framebuffer};
-pub use generated::*;
 use mapping::StrGlyphMapping;
-pub use mono_image::MonoImage;
+pub use mono_image::{MonoImage, MonoImageStack, MonoRleImage};
 pub use multi_mono_text_style::{
     MultiMonoLineHeight, MultiMonoTextStyle, MultiMonoTextStyleBuilder,
 };
@@ -24,24 +20,75 @@ pub use static_text::StaticText;
 
 pub type MultiMonoFontList<'a> = &'a [&'a MultiMonoFont<'a>];
 
-use embedded_graphics::{
-    geometry::{OriginDimensions, Point},
-    image::ImageRaw,
-    pixelcolor::BinaryColor,
-    primitives::Rectangle,
-};
-use sub_image::SubImage;
+cfg_if::cfg_if! {
+    if #[cfg(feature = "font-rawimg")] {
+        mod generated;
+        mod sub_image;
 
-#[cfg(not(feature = "big-character-size"))]
-pub type ChSzTy = u8;
-#[cfg(feature = "big-character-size")]
-pub type ChSzTy = u16;
+        pub use generated::*;
+        use sub_image::SubImage;
+        use embedded_graphics::{
+            geometry::{OriginDimensions, Point},
+            image::ImageRaw,
+            pixelcolor::BinaryColor,
+            primitives::Rectangle,
+        };
+    }
+}
+
+#[cfg(not(any(feature = "font-rawimg", feature = "font-rle")))]
+compile_error!("At least one of the features 'font-rawimg' or 'font-rle' must be enabled");
+
+cfg_if::cfg_if! {
+    if #[cfg(feature = "big-character-size")] {
+        pub type ChSzTy = u16;
+    } else {
+        pub type ChSzTy = u8;
+    }
+}
+
+cfg_if::cfg_if! {
+    if #[cfg(feature = "rle-index-ty-32bit")] {
+        pub type RleIdxTy = u32;
+    } else {
+        pub type RleIdxTy = u16;
+    }
+}
 
 /// A trait for objects that can be scaled.
 pub trait Scalable {
     type T;
     fn set_scale(&mut self, scale: Self::T);
     fn get_scale(&self) -> Self::T;
+}
+
+cfg_if::cfg_if! {
+    if #[cfg(feature = "font-rle")] {
+        #[derive(Clone, Copy, PartialEq)]
+        pub struct RLERaw {
+            glyphs_data: &'static [u8],
+            glyphs_index: &'static [RleIdxTy],
+        }
+
+        impl RLERaw {
+            pub const fn new(glyphs_data: &'static [u8], glyphs_index: &'static [RleIdxTy]) -> Self {
+                Self {
+                    glyphs_data,
+                    glyphs_index,
+                }
+            }
+        }
+    }
+}
+
+#[derive(Clone, Copy, PartialEq)]
+pub enum GlyphData {
+    /// Raw image data.
+    #[cfg(feature = "font-rawimg")]
+    ImgRaw(ImageRaw<'static, BinaryColor>),
+    /// RLE compressed image data.
+    #[cfg(feature = "font-rle")]
+    RLE(RLERaw),
 }
 
 /// Monospaced bitmap font.
@@ -51,8 +98,8 @@ pub trait Scalable {
 /// [module documentation]: self
 #[derive(Clone, Copy)]
 pub struct MultiMonoFont<'a> {
-    /// Raw image data containing the font.
-    pub image: ImageRaw<'a, BinaryColor>,
+    /// Font data.
+    pub glyph_data: GlyphData,
 
     /// Size of a single character in pixel.
     pub character_size: CharSize,
@@ -73,15 +120,17 @@ pub struct MultiMonoFont<'a> {
 }
 
 impl MultiMonoFont<'_> {
-    /// Returns a subimage for a glyph.
-    pub(crate) fn glyph(&self, c: char) -> SubImage<'_, ImageRaw<'_, BinaryColor>> {
-        if self.character_size.width == 0
-            || self.image.size().width < self.character_size.width as u32
-        {
-            return SubImage::new_unchecked(&self.image, Rectangle::zero());
+    #[cfg(feature = "font-rawimg")]
+    pub(crate) fn glyph_img<'b>(
+        &self,
+        c: char,
+        image: &'b ImageRaw<'static, BinaryColor>,
+    ) -> SubImage<'b, ImageRaw<'_, BinaryColor>> {
+        if self.character_size.width == 0 || image.size().width < self.character_size.width as u32 {
+            return SubImage::new_unchecked(image, Rectangle::zero());
         }
 
-        let glyphs_per_row = self.image.size().width / self.character_size.width as u32;
+        let glyphs_per_row = image.size().width / self.character_size.width as u32;
 
         // Char _code_ offset from first char, most often a space
         // E.g. first char = ' ' (32), target char = '!' (33), offset = 33 - 32 = 1
@@ -93,19 +142,34 @@ impl MultiMonoFont<'_> {
         let char_y = row * self.character_size.height as u32;
 
         SubImage::new_unchecked(
-            &self.image,
+            image,
             Rectangle::new(
                 Point::new(char_x as i32, char_y as i32),
                 self.character_size.size(),
             ),
         )
     }
+
+    #[cfg(feature = "font-rle")]
+    pub(crate) fn glyph_rle(&self, c: char, rle_raw: &RLERaw) -> crate::glyph_reader::GlyphReader {
+        let idx = self.glyph_mapping.index(c);
+        let glyph = if idx == 0 {
+            rle_raw.glyphs_data
+        } else {
+            let start = rle_raw.glyphs_index.get(idx - 1).copied().unwrap_or(0) as usize;
+            rle_raw
+                .glyphs_data
+                .get(start..)
+                .unwrap_or(rle_raw.glyphs_data)
+        };
+        crate::glyph_reader::GlyphReader::new(glyph)
+    }
 }
 
 impl PartialEq for MultiMonoFont<'_> {
     #[allow(trivial_casts)]
     fn eq(&self, other: &Self) -> bool {
-        self.image == other.image
+        self.glyph_data == other.glyph_data
             && self.character_size == other.character_size
             && self.character_spacing == other.character_spacing
             && self.baseline == other.baseline
@@ -113,10 +177,20 @@ impl PartialEq for MultiMonoFont<'_> {
     }
 }
 
-impl fmt::Debug for MultiMonoFont<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+impl core::fmt::Debug for GlyphData {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        match self {
+            #[cfg(feature = "font-rawimg")]
+            GlyphData::ImgRaw(_) => f.write_str("ImageRaw"),
+            #[cfg(feature = "font-rle")]
+            GlyphData::RLE(_) => f.write_str("RLERaw"),
+        }
+    }
+}
+impl core::fmt::Debug for MultiMonoFont<'_> {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         f.debug_struct("MultiMonoFont")
-            .field("image", &self.image)
+            .field("glyph_data", &self.glyph_data)
             .field("character_size", &self.character_size)
             .field("character_spacing", &self.character_spacing)
             .field("baseline", &self.baseline)
@@ -125,27 +199,22 @@ impl fmt::Debug for MultiMonoFont<'_> {
     }
 }
 
-#[cfg(feature = "defmt")]
-impl ::defmt::Format for MultiMonoFont<'_> {
-    fn format(&self, f: ::defmt::Formatter) {
-        ::defmt::write!(
-            f,
-            "MultiMonoFont {{ image: {}, character_size: {}, character_spacing: {}, baseline: {}, strikethrough: {}, underline: {}, .. }}",
-            &self.image,
-            &self.character_size,
-            &self.character_spacing,
-            &self.baseline,
-            &self.strikethrough,
-            &self.underline,
-
-        )
+cfg_if::cfg_if! {
+    if #[cfg(feature = "font-rawimg")] {
+        const NULL_FONT: MultiMonoFont = MultiMonoFont {
+            glyph_data: GlyphData::ImgRaw(ImageRaw::new(&[], 1)),
+            character_size: CharSize::zero(),
+            character_spacing: 0,
+            baseline: 0,
+            glyph_mapping: &StrGlyphMapping::new("", 0),
+        };
+    } else {
+        const NULL_FONT: MultiMonoFont = MultiMonoFont {
+            glyph_data: GlyphData::RLE(RLERaw::new(&[], &[])),
+            character_size: CharSize::zero(),
+            character_spacing: 0,
+            baseline: 0,
+            glyph_mapping: &StrGlyphMapping::new("", 0),
+        };
     }
 }
-
-const NULL_FONT: MultiMonoFont = MultiMonoFont {
-    image: ImageRaw::new(&[], 1),
-    character_size: CharSize::zero(),
-    character_spacing: 0,
-    baseline: 0,
-    glyph_mapping: &StrGlyphMapping::new("", 0),
-};